+++ pycantonese/tests/test_version_number.py
@@ -10,7 +10,9 @@ _REPO_DIR = os.path.dirname(os.path.dirname(_THIS_DIR))
 
 def test_version_number_match_with_changelog():
     """__version__ and CHANGELOG.md match for the latest version number."""
-    changelog = open(os.path.join(_REPO_DIR, "CHANGELOG.md")).read()
+    changelog = open(
+        os.path.join(_REPO_DIR, "CHANGELOG.md"), encoding="utf-8"
+    ).read()
     # latest version number in changelog = the 1st occurrence of '[x.y.z]'
     version_in_changelog = (
         re.search(r"\[\d+\.\d+\.\d+\]", changelog).group().strip("[]")