+++ pycantonese/tests/test_word_segmentation.py
@@ -1,0 +1,5 @@
+from pycantonese.word_segmentation import segment
+
+
+def test_segment():
+    assert segment("廣東話容唔容易學？") == ["廣東話", "容", "唔", "容易", "學", "？"]